feat(sql): Reintroduce ISqlService interface with methods for managing SQL servers and databases; add SqlAuthorizationException for error handling
